整理 srt cop 代码
